Refactoring: removed parsing interfaces; parsing params declared in constants; create some specific base classes; added Directives group
