+++ docs/ru/http-headers.md
@@ -24,57 +24,7 @@ ETag: W/"3d19ee-418a3-5a0d6b89d613d;5a0f5e1acb07d"
 Content-Disposition: attachment; filename="filename.jpg"
 ```
 
-С целью выделения общих правил парсинга в контексте пакета `yiisoft/http` HTTP-заголовки разбиты на несколько частично
-пересекающихся групп:
-
-1. Заголовки, имеющие одно значение (Date, ETag) или несколько, но в особом формате (Content-Range).
-2. Заголовки, которые позволяют иметь множество значений (список). Значения таких заголовков могут быть перечислены в
-   одном заголовке через запятую или в нескольких одноимённых заголовках. Это такие заголовки, как Accept, Allow,
-   Keep-Alive и т.д.
-3. Заголовки, которые могут иметь параметры. Параметры отделяются символом ";" от значения. В примере выше это заголовки
-   Accept, Keep-Alive и Content-Disposition.
-4. Группа заголовков со множественными значениями и параметрами, среди которых присутствует особый параметр "q"
-   (quality), который может означает приоритет, "качество" или "вес" значения. В приведённом примере это заголовок
-   Allow.
-
-Правила обработки каждого заголовка описываются в отдельных классах, наследующихся от класса
-`\Yiisoft\Http\Header\Value\BaseHeaderValue`.
-Для того, чтобы отнести заголовок к той или иной группе из списка выше, в описываемом заголовок классе следует указать
-интерфейс, соответствующий этой группе.
-По умолчанию, если класс заголовка не имеет интерфейсов, задающих правила парсинга, то он обрабатывается как заголовок
-первой группы.
-
-Пример:
-
-```php
-use \Yiisoft\Http\Header\Rule\ListedValues;
-use \Yiisoft\Http\Header\Rule\WithParams;
-use \Yiisoft\Http\Header\Value\BaseHeaderValue;
-
-// В поле значения заголовка передаётся только одно значение
-final class Date extends BaseHeaderValue {}
-// Заголовок со списком значений
-final class Allow extends BaseHeaderValue implements ListedValues {}
-// Со списком значений и параметрами
-final class Forwarded extends BaseHeaderValue implements WithParams, ListedValues {}
-```
-
-Указание интерфейсов влияет только на процессы парсинга и генерирования поля значения заголовка. Таким образом, при
-отсутствующем интерфейсе `WithParams` вам будут доступны методы `withParams()` и `getParams()`, но при генерировании
-строки параметры будут игнорироваться, если метод генератора не переопределён.
-
-Рассмотрим поведение парсера на примере произвольного заголовка `X-Header`:
-
-```php
-$header = XHeader::createHeader()->add('foo;param1=test;q=0.5, bar');
-```
- - Если класс `XHeader` не будет иметь интерфейс `ListedValues`, то вся строка, передаваемая в метод `add()` станет
-   единственным значением заголовка.
- - При наличии только одного интерфейса `ListedValues` у заголовка будет два значения: `foo;param1=test;q=0.5` и `bar`.
- - `ListedValues` и `WithParams` вместе дадут ожидаемый разбор на значения и параметры, однако параметр `q` не будет
-   учитываться в качестве параметра сортировки, а метод `getQuality()` будет возвращать "1", как значение по умолчанию.
- - Указание интерфейса `WithQualityParam` добавит автоматическую сортировку значений, а метод `getQuality()` будет
-   зависеть от параметра `q`.
+Все заголовки разные и со своими особенностями...
 
 ## Класс Header
