--- src/Header/Rule/ListedValues.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Yiisoft\Http\Header\Rule;
-
-interface ListedValues
-{
-}
