--- src/Header/Rule/WithParams.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Yiisoft\Http\Header\Rule;
-
-interface WithParams
-{
-    /**
-     * @return array<string, string>
-     */
-    public function getParams(): array;
-    /**
-     * @param array<string, string> $params
-     * @return $this
-     */
-    public function withParams(array $params);
-}
