--- src/Header/Rule/WithQualityParam.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace Yiisoft\Http\Header\Rule;
-
-/**
- * @link https://tools.ietf.org/html/rfc7231#section-5.3.1
- */
-interface WithQualityParam extends ListedValues, WithParams
-{
-    /**
-     * @return string value between 0.000 and 1.000
-     */
-    public function getQuality(): string;
-}
