+++ src/Header/Value/Allow.php
@@ -4,12 +4,14 @@ declare(strict_types=1);
 
 namespace Yiisoft\Http\Header\Value;
 
-use Yiisoft\Http\Header\Rule\ListedValues;
+use Yiisoft\Http\Header\Internal\BaseHeaderValue;
 
 /**
  * @link https://tools.ietf.org/html/rfc7231#section-7.4.1
  */
-final class Allow extends BaseHeaderValue implements ListedValues
+final class Allow extends BaseHeaderValue
 {
     public const NAME = 'Allow';
+
+    protected const PARSING_LIST = true;
 }
