+++ src/Header/Value/Cache/Age.php
@@ -4,9 +4,8 @@ declare(strict_types=1);
 
 namespace Yiisoft\Http\Header\Value\Cache;
 
-use InvalidArgumentException;
 use Yiisoft\Http\Header\ParsingException;
-use Yiisoft\Http\Header\Value\BaseHeaderValue;
+use Yiisoft\Http\Header\Internal\BaseHeaderValue;
 
 /**
  * @link https://tools.ietf.org/html/rfc7234#section-5.1
