+++ src/Header/Value/Connection.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace Yiisoft\Http\Header\Value;
 
-use Yiisoft\Http\Header\Rule\ListedValues;
+use Yiisoft\Http\Header\Internal\BaseHeaderValue;
 
-class Connection extends BaseHeaderValue implements ListedValues
+class Connection extends BaseHeaderValue
 {
     public const NAME = 'Connection';
+
+    protected const PARSING_LIST = true;
 }
