+++ src/Header/Value/DirectiveValue.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Http\Header\Value;
+
+use Yiisoft\Http\Header\DirectiveHeader;
+use Yiisoft\Http\Header\Internal\DirectivesHeaderValue;
+
+final class DirectiveValue extends DirectivesHeaderValue
+{
+    public static function createHeader(string $headerName = 'Undefined'): DirectiveHeader
+    {
+        return new DirectiveHeader($headerName);
+    }
+}
