+++ src/Header/Value/SimpleValue.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Http\Header\Value;
 
 use Yiisoft\Http\Header\Header;
+use Yiisoft\Http\Header\Internal\BaseHeaderValue;
 
 final class SimpleValue extends BaseHeaderValue
 {
