+++ tests/Header/HeaderTest.php
@@ -6,7 +6,7 @@ use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use Yiisoft\Http\Header\Header;
 use Yiisoft\Http\Header\Value\Accept\Accept;
-use Yiisoft\Http\Header\Value\BaseHeaderValue;
+use Yiisoft\Http\Header\Internal\BaseHeaderValue;
 use Yiisoft\Http\Header\Value\Date;
 use Yiisoft\Http\Header\Value\SimpleValue;
 use Yiisoft\Http\Tests\Header\Value\Stub\DummyHeaderValue;
