+++ tests/Header/Value/DateTest.php
@@ -24,10 +24,10 @@ class DateTest extends TestCase
     }
     public function testToString()
     {
-        $dateStr = '2020-01-01 00:00:00 +0000';
+        $dateStr = 'Friday, 04-Jul-08 08:42:36 GMT';
         $value = (new Date())->withValue($dateStr);
 
-        $this->assertSame('Wed, 01 Jan 2020 00:00:00 GMT', (string)$value);
+        $this->assertSame('Fri, 04 Jul 2008 08:42:36 GMT', (string)$value);
     }
     public function testToStringIncorrect()
     {
@@ -39,10 +39,26 @@ class DateTest extends TestCase
     }
     public function testGetDatetimeValue()
     {
-        $dateStr = '2020-01-01 00:00:00 +0000';
+        $dateStr = 'Fri Jul 4 08:42:36 2008';
         $value = new Date($dateStr);
 
         $this->assertInstanceOf(\DateTimeImmutable::class, $value->getDatetimeValue());
+    }
+    public function testRFC7231()
+    {
+        $dateStr = 'Fri, 04 Jul 2008 08:42:36 GMT';
+        $value = new Date($dateStr);
+
+        $this->assertInstanceOf(\DateTimeImmutable::class, $value->getDatetimeValue());
+        $this->assertSame('Fri, 04 Jul 2008 08:42:36 GMT', (string)$value);
+    }
+    public function testRFC850()
+    {
+        $dateStr = 'Friday, 04-Jul-08 08:42:36 GMT';
+        $value = new Date($dateStr);
+
+        $this->assertInstanceOf(\DateTimeImmutable::class, $value->getDatetimeValue());
+        $this->assertSame('Fri, 04 Jul 2008 08:42:36 GMT', (string)$value);
     }
     public function testConstructWithDatetimeInterface()
     {
@@ -62,5 +78,13 @@ class DateTest extends TestCase
         $this->assertTrue($value->hasError());
         $this->assertNull($value->getDatetimeValue());
         $this->assertSame('not a date', (string)$value);
+    }
+    public function testGetDatetimeValueFromIncorrectForHttp()
+    {
+        $value = (new Date())->withValue('2008-07-12 10:15');
+
+        $this->assertTrue($value->hasError());
+        $this->assertNull($value->getDatetimeValue());
+        $this->assertSame('2008-07-12 10:15', (string)$value);
     }
 }
