+++ tests/Header/Value/Stub/DummyHeaderValue.php
@@ -2,7 +2,7 @@
 
 namespace Yiisoft\Http\Tests\Header\Value\Stub;
 
-use Yiisoft\Http\Header\Value\BaseHeaderValue;
+use Yiisoft\Http\Header\Internal\BaseHeaderValue;
 
 final class DummyHeaderValue extends BaseHeaderValue
 {
