+++ tests/Header/Value/Stub/SortedHeaderValue.php
@@ -2,11 +2,15 @@
 
 namespace Yiisoft\Http\Tests\Header\Value\Stub;
 
-use Yiisoft\Http\Header\Rule\WithQualityParam;
+use Yiisoft\Http\Header\Internal\WithParamsHeaderValue;
 
-final class SortedHeaderValue extends \Yiisoft\Http\Header\Value\BaseHeaderValue implements WithQualityParam
+final class SortedHeaderValue extends WithParamsHeaderValue
 {
     public const NAME = 'Test-Quality';
+
+    protected const PARSING_LIST = true;
+    protected const PARSING_Q_PARAM = true;
+
     public function setQuality(string $q): bool
     {
         return parent::setQuality($q);
